Implement a stand-alone benchmarking script:
  * Script simulates one of several data volumes, chosen based
    on runtime job size.

  * Ground scheduler moved into the toast package to support calling
    directly at the job start without using subprocess.

  * Several small build system cleanups.
